git commit -m "Changed license to MIT.
Areas of this project are in part due to other applications/projects and therefore
it should not be required to use GNU license if using code in this project. Instead,
uses of this project should seek out requirements and licenses in any projects
used here for use in applications/projects.